Make the animation frame rate a floating-point value
The per-animation frame rate (the .wxmx "fr" attribute) and the default
framerate in the configuration dialog were integers, so animations could only
run at whole frames per second and never below 1 fps. Make the whole path
double-based (AnimationCell::Get/SetFrameRate, m_framerate, the constructors,
Configuration::DefaultFramerate, the config-dialog control) so fractional and
slow (< 1 fps) animations are possible. The lower clamp is dropped from 1 to
0.05 fps; the timer interval and GIF frame delay are computed as
1000/framerate ms as before.

- The "fr" attribute is now parsed with ToDouble (which still reads the integer
  values written by older wxMaxima) and written with %g.
- The config dialog uses a wxSpinCtrlDouble (2 decimals) instead of wxSpinCtrl.
- The config key is renamed "DefaultFramerate" -> "defaultFramerateFPS" so the
  int->double type change does not trip the wxRegConfig (Windows) / macOS
  type-mismatch assert when an old integer-typed value is read back as a double;
  the old key is left orphaned and the default (12 fps) applies once.

NEWS updated.

+++ src/Configuration.h
@@ -1077,8 +1077,8 @@ public:
   int DefaultPlotHeight() const {return m_defaultPlotHeight;}
   void DefaultPlotWidth(int px){m_defaultPlotWidth = px;}
   int DefaultPlotWidth() const {return m_defaultPlotWidth;}
-  void DefaultFramerate(int fps){m_defaultFramerate = fps;}
-  int DefaultFramerate() const {return m_defaultFramerate;}
+  void DefaultFramerate(double fps){m_defaultFramerate = fps;}
+  double DefaultFramerate() const {return m_defaultFramerate;}
   void TocDepth(int depth){m_tocDepth = depth;}
   int TocDepth() const {return m_tocDepth;}
   bool TeXExponentsAfterSubscript() const {return m_TeXExponentsAfterSubscript;}
@@ -1366,7 +1366,7 @@ private:
   Appearance m_appearance = Appearance::followSystem;
   long m_recentItems;
   int m_bitmapScale;
-  int m_defaultFramerate;
+  double m_defaultFramerate;
   int m_tocDepth;
   int m_maxClipbrd_BitmapMegabytes;
   int m_autoSaveMinutes;
